Safari-611.1.10/LayoutTests/webanimations/js-wrapper-kept-alive-expected.txt

This test checks that registering an event listener on an animation whose JS wrapper would otherwise be garbage-collected still fires registered event listeners.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.target._isMyAnimation is true
PASS successfullyParsed is true

TEST COMPLETE
